face_recognition/requirements.txt
numpy==1.16.4
Pillow==6.1.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
